requirements.txt
pandas
numpy
nltk
pybliometrics
markdown2
weasyprint
python-dotenv
anthropic
voyageai
ftfy
scikit-learn
umap-learn
bertopic
